backend/requirements.txt
Flask==3.0.0
gunicorn==21.0.1
stanza==1.6.1
google-cloud-storage==2.13.0